fix: 각주 히트를 페인트 계층으로 되돌린다 (#94)
리뷰(R41) 두 건. 뿌리가 하나다 — **히트 테스터가 `HwpBlockContentWalker`를
쓰지 않고** 순서와 영역을 손으로 다시 구현했다. 페인트·선택은 walker가 단일
소유하는데 히트만 갈라져 나와 두 방향으로 어긋났다.

① 겹치는 개체에서 덮인 URL이 열렸다 (R41 #1)

히트가 문단 → 글상자(저장 순서) → 표 순이었다. 페인트는 글 뒤로 개체 → 문단
텍스트 → 나머지 개체 → 안쪽 표 순이고 개체는 zOrder로 정렬되므로, 히트는 그
**역순**이어야 위에 그려진 것이 이긴다. `.inFrontOfText` 글상자가 링크 있는
각주 텍스트를 덮으면 보이지 않는 아래 URL이 열렸다.
정렬은 `HwpBlockContentWalker.footnoteTextboxesInPaintOrder`가 준다 — 히트가
정렬을 자체 구현하면 같은 어긋남이 반복되므로 소유권을 walker에 남겼다.

② 자손 오버플로 링크가 안 눌렸다 (R41 #2)

두 겹에서 막혀 있었다. (a) `hitEligibleFrame`이 최상위 개체 rect만 union해,
각주 안 표의 셀·글상자 안 문단이 자기 컨테이너를 넘어 그려지면 그 띠가 자격
영역 밖이었다. (b) 통과하더라도 `where nested.rect.contains` / `textbox.rect.
contains` 게이트가 조회 전에 기각했다.
(a)는 `walkFootnote` 방문 rect를 그대로 union해 페인트와 같은 재귀로 맞췄고,
(b)는 게이트를 없애 포함 판정을 자손 rect를 아는 안쪽 함수에 맡겼다.
R39 #3이 한 겹만 열어 둔 것을 마저 연 것이다 (회귀가 아니라 미완성이었다).

가드 (revert 시 리뷰가 기술한 증상 그대로 실패 확인):
- `testOverlappingFootnoteObjectHitFollowsReversePaintOrder` — 글 앞으로/글
  뒤로 두 배치를 모두 태운다. revert 시 `overlay` 대신 `covered`가 열린다
- `testFootnoteDescendantOverflowHyperlinkIsTappable` — 셀(500pt)이 표
  rect(100pt)를 넘는 형상. revert 시 nil (안 눌림)

`HwpFootnoteObjectLayoutTests`가 file_length 상한(700)을 넘어 하이퍼링크·히트
확장을 `HwpFootnoteHyperlinkHitTests.swift`로 분리했다 (454 + 294줄).

검증: `HWP_SNAPSHOT_TESTS=1 swift test` — 렌더 픽셀 해시가 **양 폰트 모드 모두**
전 픽스처 × 전 페이지 무변화(54.3s/60.0s 실제 실행), fidelity·블록 스냅샷·골든·
페이지 수 무변화. 전수 1,717개 0 실패. 히트 경로만 바뀌어 페인트는 무영향이다.

표 셀 경로(`tableHyperlinkURL`)에도 같은 순서 문제가 있지만 이 PR 이전부터
있던 것이라 건드리지 않았다.
